fix: move PollNotebookLogsForStatus to shared support to fix depguard violations
The depguard boundary rules added in 1a3736d prevent cross-suite imports,
but tests/odh/ was still importing from tests/trainer/utils/. Move
PollNotebookLogsForStatus to tests/common/support/core.go and update all
callers.
