+++ tests/trainer/utils/utils_notebook.go
@@ -18,8 +18,6 @@ package trainer
 
 import (
 	"fmt"
-	"strings"
-	"time"
 
 	. "github.com/onsi/gomega"
 
@@ -46,36 +44,4 @@ func WaitForNotebookPodRunning(test Test, namespace string) (string, string) {
 
 	pods := GetPods(test, namespace, metav1.ListOptions{LabelSelector: labelSelector, FieldSelector: "status.phase=Running"})
 	return pods[0].Name, pods[0].Spec.Containers[0].Name
-}
-
-// PollNotebookLogsForStatus polls the notebook container logs until a definitive NOTEBOOK_STATUS line appears or timeout.
-func PollNotebookLogsForStatus(test Test, namespace, podName, containerName string, timeout time.Duration) error {
-	var finalErr error
-
-	// Tail last N lines similar to the previous kubectl --tail
-	var tail int64 = 2000
-	getLogs := PodLog(test, namespace, podName, corev1.PodLogOptions{
-		Container: containerName,
-		TailLines: &tail,
-	})
-
-	// Track failure signal while polling
-	sawFailure := false
-	test.Eventually(func() bool {
-		logs := getLogs(test)
-		switch {
-		case strings.Contains(logs, "NOTEBOOK_STATUS: SUCCESS"):
-			return true
-		case strings.Contains(logs, "NOTEBOOK_STATUS: FAILURE"):
-			sawFailure = true
-			return true
-		default:
-			return false
-		}
-	}, timeout).Should(BeTrue(), "Notebook did not reach definitive state")
-
-	if sawFailure {
-		finalErr = fmt.Errorf("Notebook execution failed")
-	}
-	return finalErr
 }
